% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters.R
\name{build_rosters}
\alias{build_rosters}
\title{Main function for building roster files}
\usage{
build_rosters(season = nflreadr:::most_recent_season(roster = TRUE))
}
\arguments{
\item{season}{season to build, defaults to nflreadr::most_recent_season(roster = TRUE)}
}
\description{
Main function for building roster files
}
