% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters_shield.R
\name{build_rosters_shieldapi}
\alias{build_rosters_shieldapi}
\title{Build rosters from shield api}
\usage{
build_rosters_shieldapi(season)
}
\arguments{
\item{season}{integer: season to build}
}
\value{
roster data frame for season
}
\description{
Build rosters from shield api
}
