% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters_ngs.R
\name{build_rosters_weekly_ngsapi}
\alias{build_rosters_weekly_ngsapi}
\title{Build weekly rosters from NGS API}
\usage{
build_rosters_weekly_ngsapi(season)
}
\arguments{
\item{season}{integer: season to build}
}
\value{
dataframe of weekly rosters
}
\description{
Build weekly rosters from NGS API
}
