% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_roster_ids.R
\name{fill_ids}
\alias{fill_ids}
\title{Fill and clean IDs
This function applies ID cleaning and processing from
\code{nflreadr::load_players()} and \code{nflreadr::load_ff_playerids()}}
\usage{
fill_ids(roster)
}
\arguments{
\item{roster}{dataframe of roster info}
}
\value{
dataframe with filled player ids
}
\description{
Fill and clean IDs
This function applies ID cleaning and processing from
\code{nflreadr::load_players()} and \code{nflreadr::load_ff_playerids()}
}
