% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosters_players.R
\name{load_nflverse_players}
\alias{load_nflverse_players}
\title{Process \code{load_players()} as root of players dataframe}
\usage{
load_nflverse_players()
}
\value{
dataframe of \code{load_players()} output
}
\description{
Process \code{load_players()} as root of players dataframe
}
